+++ .gitignore
@@ -4,3 +4,4 @@
 # IntelliJ
 out/
 target/
+*.iml
