--- core/src/main/java/by/itacademy/service/exception/UserExistsException.java
@@ -1,7 +1,0 @@
-package by.itacademy.service.exception;
-
-public class UserExistsException extends RuntimeException{
-    public UserExistsException(String login) {
-        super(login);
-    }
-}
