--- core/src/main/java/by/itacademy/service/exception/UserNotFindException.java
@@ -1,7 +1,0 @@
-package by.itacademy.service.exception;
-
-public class UserNotFindException extends RuntimeException{
-    public UserNotFindException(String login) {
-        super(login);
-    }
-}
