--- core/src/main/java/by/itacademy/service/exception/UserPasswordIncorrectException.java
@@ -1,7 +1,0 @@
-package by.itacademy.service.exception;
-
-public class UserPasswordIncorrectException extends RuntimeException{
-    public UserPasswordIncorrectException(String login) {
-        super(login);
-    }
-}
