--- core/src/main/java/by/itacademy/service/util/PasswordDecoder.java
@@ -1,5 +1,0 @@
-package by.itacademy.service.util;
-
-public interface PasswordDecoder {
-    String decode(String password);
-}
