--- core/src/main/java/by/itacademy/service/util/PasswordEncoder.java
@@ -1,5 +1,0 @@
-package by.itacademy.service.util;
-
-public interface PasswordEncoder {
-    String encode(String password);
-}
