+++ domain/src/main/java/dev/akyl/academy/domain/permission/adapter/PermissionRepository.java
@@ -1,0 +1,4 @@
+package dev.akyl.academy.domain.permission.adapter;
+
+public interface PermissionRepository {
+}
