+++ domain/src/main/java/dev/akyl/academy/domain/role/adapter/RoleRepository.java
@@ -1,0 +1,7 @@
+package dev.akyl.academy.domain.role.adapter;
+
+import dev.akyl.academy.domain.role.Role;
+
+public interface RoleRepository {
+    Role createRole(Role role);
+}
