+++ domain/src/main/java/dev/akyl/academy/domain/user/adapter/UserPasswordEncryption.java
@@ -1,0 +1,6 @@
+package dev.akyl.academy.domain.user.adapter;
+
+public interface UserPasswordEncryption {
+    String decode(String password);
+    String encode(String password);
+}
