+++ domain/src/main/java/dev/akyl/academy/domain/user/exception/UserExistsException.java
@@ -1,0 +1,7 @@
+package dev.akyl.academy.domain.user.exception;
+
+public class UserExistsException extends RuntimeException{
+    public UserExistsException(String login) {
+        super(login);
+    }
+}
