+++ domain/src/main/java/dev/akyl/academy/domain/user/exception/UserNotFindException.java
@@ -1,0 +1,7 @@
+package dev.akyl.academy.domain.user.exception;
+
+public class UserNotFindException extends RuntimeException{
+    public UserNotFindException(String login) {
+        super(login);
+    }
+}
