+++ domain/src/main/java/dev/akyl/academy/domain/user/exception/UserPasswordIncorrectException.java
@@ -1,0 +1,7 @@
+package dev.akyl.academy.domain.user.exception;
+
+public class UserPasswordIncorrectException extends RuntimeException{
+    public UserPasswordIncorrectException(String login) {
+        super(login);
+    }
+}
