+++ infrastructure/pom.xml
@@ -9,23 +9,11 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>app</artifactId>
+    <artifactId>infrastructure</artifactId>
     <dependencies>
         <dependency>
             <groupId>by.itacademy</groupId>
-            <artifactId>infrastructure</artifactId>
-            <version>1.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>by.itacademy</groupId>
             <artifactId>domain</artifactId>
-            <version>1.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>by.itacademy</groupId>
-            <artifactId>conf</artifactId>
             <version>1.0-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
