--- repository/src/main/java/by/itacademy/repository/hash/SimpleDecoder.java
@@ -1,10 +1,0 @@
-package by.itacademy.repository.hash;
-
-import by.itacademy.service.util.PasswordDecoder;
-
-public class SimpleDecoder implements PasswordDecoder {
-    @Override
-    public String decode(String password) {
-        return password;
-    }
-}
