--- repository/src/main/java/by/itacademy/repository/hash/SimpleEncoder.java
@@ -1,10 +1,0 @@
-package by.itacademy.repository.hash;
-
-import by.itacademy.service.util.PasswordEncoder;
-
-public class SimpleEncoder implements PasswordEncoder {
-    @Override
-    public String encode(String password) {
-        return password;
-    }
-}
